+++ e2e/fixtures/path-config/excluded/excludedWithStrictComment.ts
@@ -1,0 +1,4 @@
+// @ts-strict
+const text: string = null;
+
+export {};