+++ src/common/constants.ts
@@ -1,2 +1,3 @@
 export const PLUGIN_NAME = 'typescript-strict-plugin';
 export const TS_STRICT_COMMENT = '@ts-strict';
+export const TS_STRICT_IGNORE_COMMENT = '@ts-strict-ignore';