+++ test/cli/repository/src/notOnPath.ts
@@ -1,8 +1,0 @@
-//@ts-strict
-interface TestType {
-  bar: string;
-}
-
-const foo1: TestType | undefined = undefined;
-
-const boo1 = foo.bar;