+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2021-29-11
+
+### Changed
+- Strict by default without `@ts-strict` comment
+- Ignores file with `@ts-strict-ignore` comment
+- Migration tool `update-strict-comments` which updates comments in files which contain at least 1 strict error
+
 ## [1.1.2] - 2021-06-15
 
 ### Added